move matplotlib backend functions to _matplotlib_backend and common backend functions to _backend
