+++ nilearn/plotting/surface/tests/test_surf_plotting.py
@@ -23,9 +23,9 @@ from nilearn.plotting import (
     plot_surf_stat_map,
 )
 from nilearn.plotting.displays import PlotlySurfaceFigure, SurfaceFigure
+from nilearn.plotting.surface._matplotlib_backend import MATPLOTLIB_VIEWS
 from nilearn.plotting.surface._utils import check_surface_plotting_inputs
 from nilearn.plotting.surface.surf_plotting import (
-    MATPLOTLIB_VIEWS,
     _compute_facecolors_matplotlib,
     _get_ticks_matplotlib,
     _get_view_plot_surf_matplotlib,
@@ -642,7 +642,7 @@ def test_add_contours_line_properties(plotly, key, value, surface_image_roi):
     ],
 )
 def test_check_view_is_valid(view, is_valid):
-    from nilearn.plotting.surface.surf_plotting import _check_view_is_valid
+    from nilearn.plotting.surface._backend import _check_view_is_valid
 
     assert _check_view_is_valid(view) is is_valid
 
@@ -657,7 +657,7 @@ def test_check_view_is_valid(view, is_valid):
     ],
 )
 def test_check_hemisphere_is_valid(hemi, is_valid):
-    from nilearn.plotting.surface.surf_plotting import (
+    from nilearn.plotting.surface._backend import (
         _check_hemisphere_is_valid,
     )
 
@@ -711,7 +711,7 @@ def test_configure_title_plotly():
     ],
 )
 def test_get_bounds(data, expected):
-    from nilearn.plotting.surface.surf_plotting import _get_bounds
+    from nilearn.plotting.surface._matplotlib_backend import _get_bounds
 
     assert _get_bounds(data) == expected
     assert _get_bounds(data, vmin=0.2) == (0.2, expected[1])
